+++ core/include/i_procdef_st.h
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2019 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2019 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -28,6 +28,8 @@
 
 #ifndef __I_PROCDEF_ST_H
 #define __I_PROCDEF_ST_H
+
+#include <stdint.h>
 
 #include "instance_st.h"
 
@@ -71,7 +73,7 @@ PROCDEF ;
 
 /* System functions */
 
-typedef int SYSFUNC (INSTANCE *, int *) ;
+typedef int SYSFUNC (INSTANCE *, intptr_t *) ;
 typedef struct _sysproc
 {
 	int       code ;
