+++ modules/mod_sys/mod_sys.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright © 2006-2019 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2006-2019 SplinterGU (Fenix/Bennugd)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -47,7 +47,7 @@
 
 /* ---------------------------------------------------------------------- */
 
-static int modsys_exec( INSTANCE * my, int * params )
+static int modsys_exec( INSTANCE * my, intptr_t * params )
 {
     int mode = params[0];
     char * filename = ( char * ) string_get( params[1] );
@@ -107,7 +107,7 @@ static int modsys_exec( INSTANCE * my, int * params )
 
 /* ---------------------------------------------------------------------- */
 
-static int modsys_getenv( INSTANCE * my, int * params )
+static int modsys_getenv( INSTANCE * my, intptr_t * params )
 {
     char *e ;
     int str ;
